Split comments logic out into comment.go
