+++ pkg/schema.go
@@ -4,13 +4,10 @@ import (
 	"bytes"
 	"cmp"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"iter"
 	"maps"
 	"slices"
-	"strconv"
-	"strings"
 
 	"gopkg.in/yaml.v3"
 )
@@ -83,9 +80,9 @@ type Schema struct {
 	Description           string             `json:"description,omitempty" yaml:"description,omitempty"`
 	Comment               string             `json:"$comment,omitempty" yaml:"$comment,omitempty"`
 	ReadOnly              bool               `json:"readOnly,omitempty" yaml:"readOnly,omitempty"`
-	Default               interface{}        `json:"default,omitempty" yaml:"default,omitempty"`
+	Default               any        `json:"default,omitempty" yaml:"default,omitempty"`
 	Ref                   string             `json:"$ref,omitempty" yaml:"$ref,omitempty"`
-	Type                  interface{}        `json:"type,omitempty" yaml:"type,omitempty"`
+	Type                  any        `json:"type,omitempty" yaml:"type,omitempty"`
 	Enum                  []any              `json:"enum,omitempty" yaml:"enum,omitempty"`
 	AllOf                 []*Schema          `json:"allOf,omitempty" yaml:"allOf,omitempty"`
 	AnyOf                 []*Schema          `json:"anyOf,omitempty" yaml:"anyOf,omitempty"`
@@ -228,7 +225,7 @@ func (s *Schema) UnmarshalYAML(value *yaml.Node) error {
 }
 
 // MarshalYAML implements [yaml.Marshaler].
-func (s *Schema) MarshalYAML() (interface{}, error) {
+func (s *Schema) MarshalYAML() (any, error) {
 	switch s.Kind() {
 	case SchemaKindTrue:
 		return true, nil
@@ -260,264 +257,6 @@ func (s *Schema) SetKind(kind SchemaKind) {
 		s.kind = SchemaKindObject
 	default:
 		panic(fmt.Errorf("Schema.SetKind(%#v): unexpected kind", kind))
-	}
-}
-
-func getYAMLKind(value string) string {
-	if _, err := strconv.ParseInt(value, 10, 64); err == nil {
-		return "integer"
-	}
-	if _, err := strconv.ParseFloat(value, 64); err == nil {
-		return "number"
-	}
-	if _, err := strconv.ParseBool(value); err == nil {
-		return "boolean"
-	}
-	if value != "" {
-		return "string"
-	}
-	return "null"
-}
-
-func getSchemaURL(draft int) (string, error) {
-	switch draft {
-	case 4:
-		return "http://json-schema.org/draft-04/schema#", nil
-	case 6:
-		return "http://json-schema.org/draft-06/schema#", nil
-	case 7:
-		return "http://json-schema.org/draft-07/schema#", nil
-	case 2019:
-		return "https://json-schema.org/draft/2019-09/schema", nil
-	case 2020:
-		return "https://json-schema.org/draft/2020-12/schema", nil
-	default:
-		return "", errors.New("invalid draft version. Please use one of: 4, 6, 7, 2019, 2020")
-	}
-}
-
-func getComments(keyNode, valNode *yaml.Node) []string {
-	var comments []string
-	if keyNode != nil {
-		if keyNode.HeadComment != "" {
-			comment := keyNode.HeadComment
-			if index := strings.LastIndex(comment, "\n\n"); index != -1 {
-				// Splits after the last "comment group". In other words, given this:
-				//
-				//	# foo
-				//	# bar
-				//
-				//	# moo
-				//	# doo
-				//	hello: ""
-				//
-				// Then only consider the last "# moo" & "# doo" comments
-				comment = comment[index+2:] // +2 to get rid of the "\n\n"
-			}
-			comments = strings.Split(comment, "\n")
-		}
-		if keyNode.LineComment != "" {
-			comments = append(comments, keyNode.LineComment)
-		}
-	}
-	if valNode.LineComment != "" {
-		comments = append(comments, valNode.LineComment)
-	}
-	if keyNode != nil {
-		// Append last as they come last
-		if keyNode.FootComment != "" {
-			comments = append(comments, strings.Split(keyNode.FootComment, "\n")...)
-		}
-	}
-	return comments
-}
-
-func processList(comment string, stringsOnly bool) []interface{} {
-	comment = strings.Trim(comment, "[]")
-	items := strings.Split(comment, ",")
-
-	var list []interface{}
-	for _, item := range items {
-		trimmedItem := strings.TrimSpace(item)
-		if !stringsOnly && trimmedItem == "null" {
-			list = append(list, nil)
-		} else {
-			trimmedItem = strings.Trim(trimmedItem, "\"")
-			list = append(list, trimmedItem)
-		}
-	}
-	return list
-}
-
-func processComment(schema *Schema, commentLines []string) (isRequired, isHidden bool) {
-	isRequired = false
-	isHidden = false
-
-	for key, value := range splitCommentByParts(commentLines) {
-		switch key {
-		case "enum":
-			schema.Enum = processList(value, false)
-		case "multipleOf":
-			if v, err := strconv.ParseFloat(value, 64); err == nil {
-				if v > 0 {
-					schema.MultipleOf = &v
-				}
-			}
-		case "maximum":
-			if v, err := strconv.ParseFloat(value, 64); err == nil {
-				schema.Maximum = &v
-			}
-		case "skipProperties":
-			if v, err := strconv.ParseBool(value); err == nil && v {
-				schema.SkipProperties = true
-			}
-		case "minimum":
-			if v, err := strconv.ParseFloat(value, 64); err == nil {
-				schema.Minimum = &v
-			}
-		case "maxLength":
-			if v, err := strconv.ParseUint(value, 10, 64); err == nil {
-				schema.MaxLength = &v
-			}
-		case "minLength":
-			if v, err := strconv.ParseUint(value, 10, 64); err == nil {
-				schema.MinLength = &v
-			}
-		case "pattern":
-			schema.Pattern = value
-		case "maxItems":
-			if v, err := strconv.ParseUint(value, 10, 64); err == nil {
-				schema.MaxItems = &v
-			}
-		case "minItems":
-			if v, err := strconv.ParseUint(value, 10, 64); err == nil {
-				schema.MinItems = &v
-			}
-		case "uniqueItems":
-			if v, err := strconv.ParseBool(value); err == nil {
-				schema.UniqueItems = v
-			}
-		case "maxProperties":
-			if v, err := strconv.ParseUint(value, 10, 64); err == nil {
-				schema.MaxProperties = &v
-			}
-		case "minProperties":
-			if v, err := strconv.ParseUint(value, 10, 64); err == nil {
-				schema.MinProperties = &v
-			}
-		case "patternProperties":
-			var jsonObject map[string]*Schema
-			if err := json.Unmarshal([]byte(value), &jsonObject); err == nil {
-				schema.PatternProperties = jsonObject
-			}
-		case "required":
-			if strings.TrimSpace(value) == "true" {
-				isRequired = true
-			}
-		case "type":
-			schema.Type = processList(value, true)
-		case "title":
-			schema.Title = value
-		case "description":
-			schema.Description = value
-		case "readOnly":
-			if v, err := strconv.ParseBool(value); err == nil {
-				schema.ReadOnly = v
-			}
-		case "default":
-			var jsonObject interface{}
-			if err := json.Unmarshal([]byte(value), &jsonObject); err == nil {
-				schema.Default = jsonObject
-			}
-		case "item":
-			schema.Items = &Schema{
-				Type: value,
-			}
-		case "itemProperties":
-			if schema.Items.Type == "object" {
-				var itemProps map[string]*Schema
-				if err := json.Unmarshal([]byte(value), &itemProps); err == nil {
-					schema.Items.Properties = itemProps
-				}
-			}
-		case "itemEnum":
-			if schema.Items == nil {
-				schema.Items = &Schema{}
-			}
-			schema.Items.Enum = processList(value, false)
-		case "additionalProperties":
-			if v, err := strconv.ParseBool(value); err == nil {
-				if v {
-					schema.AdditionalProperties = &SchemaTrue
-				} else {
-					schema.AdditionalProperties = &SchemaFalse
-				}
-			}
-		case "unevaluatedProperties":
-			if v, err := strconv.ParseBool(value); err == nil {
-				schema.UnevaluatedProperties = &v
-			}
-		case "$id":
-			schema.ID = value
-		case "$ref":
-			schema.Ref = value
-		case "hidden":
-			if v, err := strconv.ParseBool(value); err == nil && v {
-				isHidden = true
-			}
-		case "allOf":
-			var jsonObject []*Schema
-			if err := json.Unmarshal([]byte(value), &jsonObject); err == nil {
-				schema.AllOf = jsonObject
-			}
-		case "anyOf":
-			var jsonObject []*Schema
-			if err := json.Unmarshal([]byte(value), &jsonObject); err == nil {
-				schema.AnyOf = jsonObject
-			}
-		case "oneOf":
-			var jsonObject []*Schema
-			if err := json.Unmarshal([]byte(value), &jsonObject); err == nil {
-				schema.OneOf = jsonObject
-			}
-		case "not":
-			var jsonObject *Schema
-			if err := json.Unmarshal([]byte(value), &jsonObject); err == nil {
-				schema.Not = jsonObject
-			}
-		}
-	}
-
-	return isRequired, isHidden
-}
-
-func splitCommentByParts(commentLines []string) iter.Seq2[string, string] {
-	return func(yield func(string, string) bool) {
-		for _, comment := range commentLines {
-			withoutPound := strings.TrimSpace(strings.TrimPrefix(comment, "#"))
-			withoutSchema, ok := strings.CutPrefix(withoutPound, "@schema")
-			if !ok {
-				continue
-			}
-			trimmed := strings.TrimSpace(withoutSchema)
-			if len(trimmed) == len(withoutSchema) {
-				// this checks if we had "# @schemafoo" instead of "# @schema foo"
-				// which works as we trimmed space before.
-				// So the check is if len("foo") == len(" foo")
-				continue
-			}
-
-			parts := strings.Split(trimmed, ";")
-			for _, part := range parts {
-				key, value, _ := strings.Cut(part, ":")
-				key = strings.TrimSpace(key)
-				value = strings.TrimSpace(value)
-
-				if !yield(key, value) {
-					return
-				}
-			}
-		}
 	}
 }
